Externalize Visitor/Transformer from ast to visitor (#141), add docs
`fluent.syntax.visitor` is matching what `@fluent/syntax` does.
The documentation covers more than just Visitor/Transformer, but
now is as good a time as any to add that.
