+++ fluent.syntax/docs/visitor.rst
@@ -1,0 +1,7 @@
+Visitor & Transformer
+=====================
+
+
+.. automodule:: fluent.syntax.visitor
+   :members:
+   :show-inheritance:
